+++ src/vuex/types.js
@@ -25,7 +25,7 @@ const mutations = {
   SET_KYC_LATEST_REQUEST: 'SET_KYC_LATEST_REQUEST',
   SET_KYC_RELATED_REQUEST: 'SET_KYC_RELATED_REQUEST',
   SET_KYC_LATEST_DATA: 'SET_KYC_LATEST_DATA',
-  SET_ACCOUNT_ROLE_RESET: 'SET_ACCOUNT_ROLE_RESET',
+  SET_ACCOUNT_ROLE_RESETED: 'SET_ACCOUNT_ROLE_RESETED',
 }
 
 const actions = {
@@ -104,7 +104,7 @@ const getters = {
   kycPreviousRequestAccountRoleToSet: 'kycPreviousRequestAccountRoleToSet',
   kycLatestData: 'kycLatestData',
   kycAvatarKey: 'kycAvatarKey',
-  isAccountRoleReset: 'isAccountRoleReset',
+  isAccountRoleReseted: 'isAccountRoleReseted',
 }
 
 export const vuexTypes = {