+++ CHANGELOG.md
@@ -208,7 +208,7 @@ for further information about branching and tagging conventions.
 - A feature that makes your tests execution fail if there is any `console.log`,
   `console.warn` or `console.error` call
 - Account reset related stuff:
-  - `isAccountRoleReset` state field to the KYC vuex module
+  - `isAccountRoleReseted` state field to the KYC vuex module
   - Checking for previous account role & resetting reason when loading
     KYC request
   - Checking for reset account role on verification pages